requirements.txt
flask
flask-cors
werkzeug
pillow
torch
torchvision
opencv-python
numpy
